fix: 스크롤 해제가 제대로 작동하지 않는 문제 해결
