+++ src/components/@common/bottomModal/BottomModal.tsx
@@ -1,5 +1,6 @@
 import React, { useEffect } from 'react';
 import { StyleBackDrop, StyleContainer } from './BottomModal.style';
+import Backdrop from './Backdrop';
 
 interface Props extends React.PropsWithChildren {
   isOpen: boolean;
@@ -7,27 +8,9 @@ interface Props extends React.PropsWithChildren {
 }
 
 const BottomModal = ({ isOpen, onClose, children }: Props) => {
-  useEffect(() => {
-    const handleKeyupEscape = (event: KeyboardEvent) => {
-      if (event.key === 'Escape') onClose();
-    };
-
-    window.addEventListener('keyup', handleKeyupEscape);
-    document.body.style.overflow = 'hidden';
-    return () => {
-      window.removeEventListener('keyup', handleKeyupEscape);
-      document.body.style.overflow = 'unset';
-    };
-  }, [onClose]);
-
   return isOpen ? (
     <>
-      <StyleBackDrop
-        className="modal-backdrop"
-        onClick={() => {
-          onClose();
-        }}
-      ></StyleBackDrop>
+      <Backdrop onClose={onClose}></Backdrop>
       <StyleContainer>{children}</StyleContainer>
     </>
   ) : null;
